% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicable_webs.R
\name{applicable_webs}
\alias{applicable_webs}
\title{Running it in cluster}
\usage{
applicable_webs(total_time, mutualism_pars)
}
\arguments{
\item{total_time}{simulation time}

\item{mutualism_pars}{muutalism-related stuff}
}
\value{
a matrix
}
\description{
Running it in cluster
}
