% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_island_mutualism.R
\name{create_island_mutualism}
\alias{create_island_mutualism}
\title{Converts simulation output into island output}
\usage{
create_island_mutualism(stt_table, simtime, island_spec, M0)
}
\arguments{
\item{stt_table}{Matrix with number of species at each time step. The first
column is "Time", followed by "nIp" (number of plant immigrants), "nAp" (number
of plant anagenesis), "nCp" (number of plant cladogenesis), "nIa" (number of
animal immigrants), "nAa" (number #' of animal anagenesis), "nCp" (number of
animal cladogenesis).}

\item{simtime}{Numeric defining the length of the simulation in time units.
For example, if an island is known to be 4 million years old, setting
simtime = 4 will simulate the entire life span of the island; setting
simtime = 2 will stop the simulation at the mid-life of the island.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}
}
\value{
list with the island information, composed stt table,
branching times of extant species, status of species on
the island and number of missing species.
}
\description{
Converts simulation output into island output
}
