% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pars.R
\name{create_mutualism_pars}
\alias{create_mutualism_pars}
\title{Title}
\usage{
create_mutualism_pars(
  lac_pars,
  mu_pars,
  K_pars,
  gam_pars,
  laa_pars,
  qgain,
  qloss,
  lambda0,
  M0,
  transprob,
  alpha
)
}
\description{
Title
}
