% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameter documentation}
\usage{
default_params_doc(
  time,
  M,
  pars,
  replicates,
  divdepmodel,
  nonoceanic_pars,
  num_guilds,
  prop_type2_pool,
  replicates_apply_type2,
  sample_freq,
  plot_sims,
  verbose,
  area_pars,
  hyper_pars,
  island_ontogeny,
  sea_level,
  extcutoff,
  shift_times,
  mainland_n,
  island_replicates,
  island_spec,
  stt_table,
  rates,
  max_rates,
  timeval,
  total_time,
  possible_event,
  maxspecID,
  mainland_spec,
  max_area,
  proportional_peak_t,
  total_island_age,
  sea_level_amplitude,
  sea_level_frequency,
  island_gradient_angle,
  d,
  x,
  simulation_outputs,
  plot_plus_one,
  type,
  plot_lists,
  ...,
  datalist,
  datatype,
  initparsopt,
  idparsopt,
  idparsfix,
  parsfix,
  idparsnoshift,
  idparsmat,
  res,
  ddmodel,
  cond,
  eqmodel,
  x_E,
  x_I,
  tol,
  maxiter,
  methode,
  optimmethod,
  CS_version,
  tolint,
  datatable,
  island_age,
  number_clade_types,
  list_type2_clades,
  epss,
  t,
  initEI,
  data_table,
  endmc,
  archipelago,
  phylo_data,
  archipelago_data,
  gam,
  laa,
  lac,
  mu,
  K,
  num_spec,
  num_immigrants,
  global_min_area_time,
  global_max_area_time,
  distance_type,
  distance_dep,
  parallel,
  cpus,
  pars1,
  pars2,
  tvec,
  initprobs,
  pb,
  island,
  title,
  plot_lists_simulations,
  plot_lists_simulations_MLE,
  kind_of_plot,
  resolution,
  resol,
  removed_timepoints,
  A,
  Amin,
  Amax,
  peak,
  proptime,
  proptime_max,
  current_area,
  jitter,
  num_cycles,
  trait_pars,
  relaxed_par,
  relaxed_rate_pars,
  brts,
  stac,
  missnumspec,
  DAISIE_par,
  DAISIE_dist_pars,
  abstolint,
  reltolint,
  pick,
  mean,
  sd,
  clado_rate,
  ext_rate,
  carr_cap,
  immig_rate,
  ana_rate
)
}
\arguments{
\item{replicates}{Integer specifying number of island replicates to be
simulated.}

\item{sample_freq}{Numeric specifing the number of units times should be
divided by for plotting purposes. Larger values will lead to plots with
higher resolution, but will also run slower.}

\item{plot_sims}{\code{Default = TRUE} It plots species-through-time (STT)
plots including total species, plant species and animal species seprately.}

\item{island_replicates}{List output from
\code{\link{DAISIE_sim_core_cr}()},
\code{\link{DAISIE_sim_core_time_dep}()},
\code{\link{DAISIE_sim_core_cr_shift}()} or
\code{\link{DAISIE_sim_min_type2}()} functions. Minimally, this must be a
list that has as many elements as replicates. Each element must be a list
with the elements \code{island_age}, \code{not_present} and \code{stt_all}.
\code{stt_all} must be a data frame with the column names \code{Time},
\code{nI}, \code{nA}, \code{nC} and \code{present}.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}

\item{stt_table}{Matrix with number of species at each time step. The first
column is "Time", followed by "nIp" (number of plant immigrants), "nAp" (number
of plant anagenesis), "nCp" (number of plant cladogenesis), "nIa" (number of
animal immigrants), "nAa" (number #' of animal anagenesis), "nCp" (number of
animal cladogenesis).}

\item{possible_event}{Numeric defining what event will happen.}

\item{simtime}{Numeric defining the length of the simulation in time units.
For example, if an island is known to be 4 million years old, setting
simtime = 4 will simulate the entire life span of the island; setting
simtime = 2 will stop the simulation at the mid-life of the island.}

\item{mutualism_pars}{A named list containing the model parameters as
created by \code{\link{create_pars}()}:
\itemize{
  \item{\code{pars[1]}: lambda^c_plant, lambda^c_animal (cladogenesis rates
  for plant and animal species respectively)}
  \item{\code{pars[2]}: mu (extinction rates for plant and animal species as
  well as mutualism related coefficient)}
  \item{\code{pars[3]}: K (carrying capacity), set K=Inf for diversity
  independence}
  \item{\code{pars[4]}: gamma (immigration rates for plant and animal species as
  well as mutualism related coefficient)}
  \item{\code{pars[5]}: lambda^a (anagenesis rate)}
  \item{\code{pars[6]}: qgain (gain links rate) each pair of species from
  different guilds has a rate to gain links if one of them is on the island
  while the other is not}
  \item{\code{pars[7]}: qloss (lose links rate) each pair of species from
  different guilds has a rate to lose links if none of them is on the island
  while they were connected before in the mainland}
  \item{\code{pars[8]}: lambda0 (coefficient mideating cospeciation)}
  \item{\code{pars[9]}: M0 (initial interaction matrix in the mainland)}
  \item{\code{pars[10]}: transprob for daughter species with this transferm
  probability inheriting links from parent species undergoing speciation,
  either anagenesis or cladogenesis.
  For example, daughter species will inherit all connnections that its partent
  species has with species from another guild if by setting transprob = 1.}
}}
}
\value{
Nothing
}
\description{
Default parameter documentation
}
