% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_pans_cmps}
\alias{get_pans_cmps}
\title{Calculating mutualistic partners and competitors}
\usage{
get_pans_cmps(Mt, status_p, status_a)
}
\arguments{
\item{Mt}{Matrix with plant species on the row and animal species on the column.
Mt is the matrix on the island having the same size as \code{M0} at \code{
timeval = 0 }, yet it constantly keeps changing as events (immigration, extinction
and speciation) happening.}

\item{status_p, status_a}{Matrix indicating species status on the island, with 0
meaning absent, 1 meaning present.}
}
\value{
A list containing the number of reciprocal partners and competitors:
  \itemize{
    \item{\code{$pans_p}: A numeric vector about the number of partners of
    plant species}
    \item{\code{$pans_a}: A numeric vector about the number of partners of
    plant species}
    \item{\code{cmps_p}: A numeric vector about the number of competitors of
    plant species}
    \item{\code{cmps_a}: A numeric vector about the number of competitors of
    animal species}
    }
}
\description{
Calculating mutualistic partners and competitors
}
