% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simulation_outputs.R
\name{get_simulation_outputs}
\alias{get_simulation_outputs}
\title{Get simulation parameters for different scenarios}
\usage{
get_simulation_outputs(
  the_path,
  params,
  effect,
  prefix,
  total_time,
  replicates
)
}
\arguments{
\item{the_path}{save the outputs}

\item{effect}{mutualsim effects, could be "none", "low", "medium", "high"}

\item{prefix}{parameters combination}

\item{total_time}{island age}

\item{replicates}{replicates}

\item{lac0}{initial cladogenesis value}

\item{mu0}{initial extinction value}

\item{K0}{initial carrying capacity}

\item{gam0}{initial immigration value}

\item{laa0}{initial anagenesis}
}
\value{
simulation outputs
}
\description{
Get simulation parameters for different scenarios
}
