% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peregrine_sim.R
\name{peregrine_sim}
\alias{peregrine_sim}
\title{Running it in cluster}
\usage{
peregrine_sim(
  total_time,
  replicates,
  mutualism_pars,
  return_parts,
  verbose = TRUE
)
}
\arguments{
\item{total_time}{simulation time}

\item{replicates}{replicates}

\item{mutualism_pars}{muutalism-related stuff}

\item{verbose}{}
}
\value{
a list
}
\description{
Running it in cluster
}
