% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_event_time.R
\name{sample_event_mutual}
\alias{sample_event_mutual}
\title{Samples what event to happen next}
\usage{
sample_event_mutual(rates)
}
\arguments{
\item{rates}{A named list of all kinds of rates as returned by
\code{\link{update_rates_mutual}}.}
}
\value{
Numeric indicating what event will happen
\itemize{
  \item{[1]: immigration event with plant species}
  \item{[2]: extinction event with plant species}
  \item{[3]: cladogenesis event with plant species}
  \item{[4]: anagenesis event with plant species}
  \item{[5]: immigration event with animal species}
  \item{[6]: extinction event with animal species}
  \item{[7]: cladogenesis event with animal species}
  \item{[8]: anagenesis event with animal species}
  \item{[9]: cospeciation event between pairs}
  \item{[10]: gain links event between pairs}
  \item{[11]: loss links event between pairs}
}
}
\description{
Samples what event to happen next
}
