% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_event_time.R
\name{sample_time_mutual}
\alias{sample_time_mutual}
\title{Samples the next time value}
\usage{
sample_time_mutual(rates, timeval)
}
\arguments{
\item{rates}{A named list of all kinds of rates as returned by
\code{\link{update_rates_mutual}}.}

\item{timeval}{Current time of simulation}
}
\value{
A named list with numeric vector containing the time of the next
timestep and the change in time.
}
\description{
Samples the next time value
}
