% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mutualism.R
\name{sim_mutualism}
\alias{sim_mutualism}
\title{sim_mutualism}
\usage{
sim_mutualism(
  total_time,
  replicates,
  mutualism_pars,
  sample_freq,
  cond_p,
  cond_a,
  return_parts = "island_parts",
  verbose
)
}
\arguments{
\item{total_time}{Numeric. Total time for the simulation.}

\item{replicates}{Integer. Number of replicates to run.}

\item{mutualism_pars}{List. Parameters used in the simulation.}

\item{sample_freq}{Numeric. Frequency of sampling data, fpr plotting.}

\item{cond_p}{Integer. Condition for plants (e.g., minimum number of present plants on the island).}

\item{cond_a}{Integer. Condition for animals (e.g., minimum number of present animals on the island).}

\item{return_parts}{Character. Specifies which part of the result to return. Options are:
- `"island_parts"` (default): Returns only the core simulation results, such as the final mutualism matrix and evolution table.
- `"additional_parts"`: Returns only additional information, such as time values, richness lists, and rate lists.}

\item{verbose}{Logical. If TRUE, prints progress updates.}
}
\value{
A list containing:
  - island_replicates: List of island replicates from simulations.
  - island_segments: Formatted data for plants and animals separated.
  - island_total: Formatted data for plants and animals combined.
}
\description{
Simulates mutualistic interactions over macro-evolutionary time with specified parameters.
}
