% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_test.R
\name{sim_test}
\alias{sim_test}
\title{Internal function of simulation}
\usage{
sim_test(total_time, mutualism_pars)
}
\value{
a named list with island information
}
\description{
Internal function of simulation
}
