% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_test_totalrates.R
\name{sim_test_totalrates}
\alias{sim_test_totalrates}
\title{For checking parameters}
\usage{
sim_test_totalrates(total_time, mutualism_pars)
}
\value{
a named list with island information (not this)
}
\description{
For checking parameters
}
