% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_state.R
\name{sim_update_state_mutualism}
\alias{sim_update_state_mutualism}
\title{Title}
\usage{
sim_update_state_mutualism(
  timeval,
  simtime,
  possible_event,
  Mt,
  status_p,
  status_a,
  maxplantID,
  maxanimalID,
  island_spec,
  stt_table,
  mutualism_pars
)
}
\description{
Title
}
