% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_states_mutual.R
\name{update_states_mutual}
\alias{update_states_mutual}
\title{Update states on the island}
\usage{
update_states_mutual(
  M0,
  Mt,
  status_p,
  status_a,
  maxplantID,
  maxanimalID,
  timeval,
  total_time,
  rates,
  possible_event,
  island_spec,
  stt_table,
  transprob
)
}
\arguments{
\item{M0}{Mainland matrix}

\item{Mt}{Island matrix}

\item{status_p}{Plant species status, with 1 present, 0 absent on island}

\item{status_a}{Animal species status, with 1 present, 0 absent on island}

\item{maxplantID}{Maximum plant species id on island}

\item{maxanimalID}{Maximum animal species id on island}

\item{timeval}{Time value during simulation, ranging from island age to the present}

\item{total_time}{Island age (total simulation time)}

\item{rates}{All kinds of rates}

\item{possible_event}{Possible event that happening on island}

\item{island_spec}{A data frame with species information on island}

\item{stt_table}{A data frame with the number of species through time}

\item{transprob}{A numerical}
}
\value{
A list
}
\description{
Update states on the island
}
