% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_island_mutual.R
\name{create_island_mutual}
\alias{create_island_mutual}
\title{Converts simulation output into island output}
\usage{
create_island_mutual(stt_table, total_time, island_spec)
}
\value{
list with the island information, composed stt table,
branching times of extant species, status of species on
the island and number of missing species.
}
\description{
Converts simulation output into island output
}
