% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stt.R
\name{plot_stt}
\alias{plot_stt}
\title{Plot species though time (stt) plots on island}
\usage{
plot_stt(several_islands, simtime)
}
\arguments{
\item{simtime}{}
}
\value{
R plot with a row of three figures.
}
\description{
plot_stt() plots three figures based on total species, plant species and animal
species on island. Each of them has the total, endemic and non-endemic with the
simulated entire time span. 2.5-97.5th percentiles are plotted in light grey,
25-75th percentiles plotted in dark grey.
}
