% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mutualism.R
\name{sim_mutualism}
\alias{sim_mutualism}
\title{sim_mutualism}
\usage{
sim_mutualism(
  total_time,
  replicates,
  mutualism_pars,
  sample_freq,
  cond_p,
  cond_a,
  verbose
)
}
\description{
sim_mutualism
}
