% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rates_mutual.R
\name{update_rates_mutual}
\alias{update_rates_mutual}
\title{Calculates algorithm rates}
\usage{
update_rates_mutual(
  M0,
  Mt,
  alphaa,
  status_p,
  status_a,
  lac_pars,
  mu_pars,
  K_pars,
  gam_pars,
  laa_pars,
  qgain,
  qloss,
  lambda0,
  transprob,
  island_spec
)
}
\value{
a named list with the updated rates.
}
\description{
Internal function that updates the all the rates at time t.
}
