% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rates_mutual.R
\name{update_rates_mutual}
\alias{update_rates_mutual}
\title{Calculates algorithm rates}
\usage{
update_rates_mutual(
  M0,
  Mt,
  status_p,
  status_a,
  lac_pars,
  mu_pars,
  K_pars,
  gam_pars,
  laa_pars,
  qgain,
  qloss,
  lambda0,
  transprob,
  island_plant,
  island_animal
)
}
\value{
a named list with the updated rates.
}
\description{
Internal function that updates the all the rates at time t.
}
\examples{
rates <- update_rates_mutualism(M0 = matrix(1, 20, ncol = 5, nrow = 4),
Mt = Mt,
status_p = matrix(0, ncol = 1, nrow = nrow(Mt)),
status_a = matrix(0, ncol = 1, nrow = ncol(Mt)),
lac_pars = c(0.5, 0.5),
mu_pars = c(0.2, 0.2, 1.0, 1.0),
K_pars = c(Inf, Inf, Inf, Inf),
gam_pars = c(0.05, 0.02),
laa_pars =  c(0.2, 0.1, 1.0, 2.0),
qgain = 1.0,
qloss = 1.0,
lambda0 = 2.0,
transprob = 1.0,
island_spec = NULL)
}
